playing with socket.io to see the potentiality
Socket.io is good to keep multiple clients in sync or create live application. In our application, we don't need either (at least, for the moment). As such, we abandon the idea of socket and we go back to HTTP, but we keep this commit to see how we structured the socket and created a sync system between vuex and server state